Grid 6.0/requirements.txt
opencv-python
tensorflow
numpy
flask
paho-mqtt
Pillow
pytesseract
open3d
